force and update options, cleanup

+++ bin/list-mods-and-orphans.sh
@@ -2,9 +2,16 @@
 # compare two directories, list modified & orphaned files 
 # TODO: list files found in second dir and not in first too
 
-[ -z "$1" -o "$1" = "-h" ] && echo "[extra='-L' constraints='-not -path whatever']" $(basename "$0") [-f] "<indir1|infile1.md5> label1 <indir2|infile2.md5> label2 out-prefix" && exit
+[ -z "$1" -o "$1" = "-h" ] && echo "[extra='-L' constraints='-not -path whatever']" $(basename "$0") [-f -u -m] "<indir1|infile1.md5> label1 <indir2|infile2.md5> label2 out-prefix" && exit
 
-[ "$1" = '-f' ] && FORCE=1 && shift
+#todo check relative path comp
+#todo: proper readarg
+
+while expr match "$1" '^-[fum]$'; do
+	[ "$1" = '-f' ] && FORCE=1 && shift
+	[ "$1" = '-u' ] && UPDATE=1 && shift
+	[ "$1" = '-m' ] && MD_ONLY=1 && shift
+done
 
 IN1="$1"
 LABEL1="$2"
@@ -19,7 +26,22 @@ expr match "${IN2}" '.*/' && IN2="${IN2%/}"
 
 make_list()
 {
-	find ${extra} "$1" -type f ${constraints} | while read f; do md5sum "$f" >> "$2"; done
+	local old=$(mktemp "$2.old.XXXXXX")
+	if [ -f "$2" ]; then
+		mv "$2" "$old"
+	fi
+
+	find ${extra} "$1" -type f ${constraints} | \
+	while read f; do
+		local m=''
+
+		[ -f "$old" ] && m=$(grep -m 1 "\ ${f}\$" "$old")
+		[ -z "$m" ] && m=$(md5sum "$f")
+		echo "$m" >> "$2"
+	done
+
+	sort -su -k 2 "$2" > "$old"
+	mv "$old" "$2"
 }
 
 
@@ -27,31 +49,35 @@ process_lists()
 {
 	rm -f "$OUT$LABEL1-only" "$OUT$LABEL1-$LABEL2-dup"
 
-	cat "$IN1" | while read md fi; do 
+	cat "$IN1" | while read md fl; do 
 		grep -q "${md}" "$IN2" && continue
-		name=$(basename "${fi}")
+		name=$(basename "${fl}")
 		loc=$(grep -m 1 "\ ${name}\$" "$IN2")
 		if [ -z "${loc}" ]; then
-			echo -e "${md}\t${fi}" >> "$OUT$LABEL1-only"
+			echo -e "${md}\t${fl}" >> "$OUT$LABEL1-only"
 		else
-			echo -e "${LABEL1}\t${fi}\t${LABEL2}\t${loc#*\ }" >> "$OUT$LABEL1-$LABEL2-dup"
+			echo -e "${LABEL1}\t${fl}\t${LABEL2}\t${loc#*\ }" >> "$OUT$LABEL1-$LABEL2-dup"
 		fi
 	done 
 }
 
-if [ -d "$IN1" ]; then
-	[ -f "$IN1.md5" -a -z "${FORCE}" ] && IN1="$IN1.md5" && break
+#todo: add new, move, replace changed (ask first option), delete removed (ask first)
 
-	rm -f "$IN1.md5"
+if [ -d "$IN1" ]; then
+	[ "$FORCE" = 1 ] && rm -f "$IN1.md5"
+	[ -f "$IN1.md5" -a -z "${UPDATE}" ] && IN1="$IN1.md5" && break
+
 	make_list "$IN1" "$IN1.md5" && IN1="$IN1.md5"
 fi
 
 if [ -d "$IN2" ]; then
-	[ -f "$IN2.md5" -a -z "${FORCE}" ] && IN2="$IN2.md5" && break
+	[ "$FORCE" = 1 ] && rm -f "$IN2.md5"
+	[ -f "$IN2.md5" -a -z "${UPDATE}" ] && IN2="$IN2.md5" && break
 
-	rm -f "$IN2.md5"
 	make_list "$IN2" "$IN2.md5" && IN2="$IN2.md5"
 fi
+
+[ -n "$MD_ONLY" ] && exit 0
 
 process_lists
